confirm is added

+++ src/components/StdDetails.jsx
@@ -1,27 +1,35 @@
-import React, { useEffect, useState,createContext } from 'react'
+import React, { useEffect, useState } from 'react'
 import { useParams, useNavigate, Link } from 'react-router-dom'
-const userContext=createContext()
 const StdDetails = () => {
   const navigate = useNavigate()
   const { id } = useParams()
   const [std, setStd] = useState({})
-  const stdUrl = `http://192.168.0.106:8080/api/Student/${id}`
-  // const stdUrl=`https://jsonplaceholder.typicode.com/users/${id}`
+  // const stdUrl = `http://192.168.0.106:8080/api/Student/${id}`
+  const stdUrl = `https://jsonplaceholder.typicode.com/users/${id}`
   const stdhandler = async (url) => {
     let response = await fetch(url)
     let resolve = await response.json()
-    setStd(resolve.Data)
+    // setStd(resolve.Data)
+    setStd(resolve) //for jsonplaceholder
   }
+
   const deleteHandler = async () => {
-    await fetch(stdUrl, { method: "DELETE" })
-    alert("Data Deleted Successfully")
-    navigate("/")
+    let result = confirm("Are you sure you want to delete this data?")
+    if (result) {
+      await fetch(stdUrl, { method: "DELETE" })
+      alert("Data Deleted Successfully")
+      navigate("/")
+    }
+    else {
+      alert("Data Not Deleted")
+      navigate("/")
+    }
   }
   useEffect(() => {
     stdhandler(stdUrl)
   }, [])
   return (
-<userContext.Provider value={std}>
+
     <div className='container3'>
       <h3>NAME:{std?.name}</h3>
       <h4>EMAIL:{std?.email}</h4>
@@ -32,8 +40,8 @@ const StdDetails = () => {
         <Link to={`/PostApi/${id}`}><button>Update</button></Link>
         <button onClick={deleteHandler}>Delete</button></div>
     </div>
-   
-    </userContext.Provider>
+
+
   )
 }
 export default StdDetails
